rename test explorer to test discovery
