+++ wasm/src/test_discovery.rs
